# Map and Set performance improvements