--- a/plottable.js
+++ b/plottable.js
@@ -126,11 +126,20 @@ var Plottable;
          */
         var Map = (function () {
             function Map() {
-                this._keyValuePairs = [];
+                if (typeof window.Map === "function") {
+                    this._es6Map = new window.Map();
+                }
+                else {
+                    this._keyValuePairs = [];
+                }
             }
             Map.prototype.set = function (key, value) {
                 if (Utils.Math.isNaN(key)) {
                     throw new Error("NaN may not be used as a key to the Map");
+                }
+                if (this._es6Map != null) {
+                    this._es6Map.set(key, value);
+                    return this;
                 }
                 for (var i = 0; i < this._keyValuePairs.length; i++) {
                     if (this._keyValuePairs[i].key === key) {
@@ -142,6 +151,9 @@ var Plottable;
                 return this;
             };
             Map.prototype.get = function (key) {
+                if (this._es6Map != null) {
+                    return this._es6Map.get(key);
+                }
                 for (var i = 0; i < this._keyValuePairs.length; i++) {
                     if (this._keyValuePairs[i].key === key) {
                         return this._keyValuePairs[i].value;
@@ -150,6 +162,9 @@ var Plottable;
                 return undefined;
             };
             Map.prototype.has = function (key) {
+                if (this._es6Map != null) {
+                    return this._es6Map.has(key);
+                }
                 for (var i = 0; i < this._keyValuePairs.length; i++) {
                     if (this._keyValuePairs[i].key === key) {
                         return true;
@@ -159,11 +174,18 @@ var Plottable;
             };
             Map.prototype.forEach = function (callbackFn, thisArg) {
                 var _this = this;
+                if (this._es6Map != null) {
+                    this._es6Map.forEach(callbackFn, thisArg);
+                    return;
+                }
                 this._keyValuePairs.forEach(function (keyValuePair) {
                     callbackFn.call(thisArg, keyValuePair.value, keyValuePair.key, _this);
                 });
             };
             Map.prototype.delete = function (key) {
+                if (this._es6Map != null) {
+                    return this._es6Map.delete(key);
+                }
                 for (var i = 0; i < this._keyValuePairs.length; i++) {
                     if (this._keyValuePairs[i].key === key) {
                         this._keyValuePairs.splice(i, 1);