--- a/plottable.js
+++ b/plottable.js
@@ -265,17 +265,16 @@ var Plottable;
              * Updates the value of the read-only parameter size
              */
             Set.prototype._updateSize = function () {
-                var newSize = 0;
                 if (this._es6Set != null) {
-                    newSize = this._es6Set.size;
+                    this.size = this._es6Set.size;
                 }
                 else {
-                    newSize = this._values.length;
+                    this.size = this._values.length;
                 }
-                Object.defineProperty(this, "size", {
-                    value: newSize,
-                    configurable: true
-                });
+                // Object.defineProperty(this, "size", {
+                //   value: newSize,
+                //   configurable: true
+                // });
             };
             return Set;
         })();