--- a/src/utils/map.ts
+++ b/src/utils/map.ts
@@ -7,12 +7,27 @@ export module Utils {
    * https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map
    */
   export class Map<K, V> {
-    private _keyValuePairs: { key: K; value: V; }[] = [];
+    private _keyValuePairs: { key: K; value: V; }[];
+    private _es6Map: any;
+
+    public constructor() {
+      if (typeof (<any>window).Map === "function") {
+        this._es6Map = new (<any>window).Map();
+      } else {
+        this._keyValuePairs = [];
+      }
+    }
 
     public set(key: K, value: V) {
       if (Utils.Math.isNaN(key)) {
         throw new Error("NaN may not be used as a key to the Map");
       }
+
+      if (this._es6Map != null) {
+        this._es6Map.set(key, value);
+        return this;
+      }
+
       for (var i = 0; i < this._keyValuePairs.length; i++) {
         if (this._keyValuePairs[i].key === key) {
           this._keyValuePairs[i].value = value;
@@ -23,7 +38,11 @@ export module Utils {
       return this;
     }
 
-    public get(key: K) {
+    public get(key: K): V {
+      if (this._es6Map != null) {
+        return this._es6Map.get(key);
+      }
+
       for (var i = 0; i < this._keyValuePairs.length; i++) {
         if (this._keyValuePairs[i].key === key) {
           return this._keyValuePairs[i].value;
@@ -32,7 +51,11 @@ export module Utils {
       return undefined;
     }
 
-    public has(key: K) {
+    public has(key: K): boolean {
+      if (this._es6Map != null) {
+        return this._es6Map.has(key);
+      }
+
       for (var i = 0; i < this._keyValuePairs.length; i++) {
         if (this._keyValuePairs[i].key === key) {
           return true;
@@ -42,12 +65,21 @@ export module Utils {
     }
 
     public forEach(callbackFn: (value: V, key: K, map: Map<K, V>) => void, thisArg?: any) {
+      if (this._es6Map != null) {
+        this._es6Map.forEach(callbackFn, thisArg);
+        return;
+      }
+
       this._keyValuePairs.forEach((keyValuePair) => {
         callbackFn.call(thisArg, keyValuePair.value, keyValuePair.key, this);
       });
     }
 
-    public delete(key: K) {
+    public delete(key: K): boolean {
+      if (this._es6Map != null) {
+        return this._es6Map.delete(key);
+      }
+
       for (var i = 0; i < this._keyValuePairs.length; i++) {
         if (this._keyValuePairs[i].key === key) {
           this._keyValuePairs.splice(i, 1);