--- a/src/utils/set.ts
+++ b/src/utils/set.ts
@@ -18,33 +18,34 @@ export module Utils {
       } else {
         this._values = [];
       }
-      this._updateSize();
+      this.size = 0;
     }
 
     public add(value: T) {
       if (this._es6Set != null) {
         this._es6Set.add(value);
-      } else {
-        if (!this.has(value)) {
-          this._values.push(value);
-        }
+        this.size = this._es6Set.size;
+        return this;
       }
 
-      this._updateSize();
+      if (!this.has(value)) {
+        this._values.push(value);
+        this.size = this._values.length;
+      }
       return this;
     }
 
     public delete(value: T): boolean {
       if (this._es6Set != null) {
         var deleted = this._es6Set.delete(value);
-        this._updateSize();
+        this.size = this._es6Set.size;
         return deleted;
       }
 
       var index = this._values.indexOf(value);
       if (index !== -1) {
         this._values.splice(index, 1);
-        this._updateSize();
+        this.size = this._values.length;
         return true;
       }
       return false;
@@ -69,20 +70,6 @@ export module Utils {
       });
     }
 
-    /**
-     * Updates the value of the read-only parameter size
-     */
-    private _updateSize() {
-      if (this._es6Set != null) {
-        this.size = this._es6Set.size;
-      } else {
-        this.size = this._values.length;
-      }
-      // Object.defineProperty(this, "size", {
-      //   value: newSize,
-      //   configurable: true
-      // });
-    }
   }
 }
 }