As part of the API refactor, we introduced the ES6 shims for Map and Set. Also, we decided not to expose the d3 map/set anymore. The problem is that our Map and Set were not efficient.

However, Chrome and FF support ES6 Map / Set for quite a while, hence we can leverage their power and cut down O(n) to O(log n) / O(1).

[Browser Support](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map):
```
Chrome: 38
Firefox: 25
IE: 11
Opera: 25
Safari: 7.1
```
For reference, [latest version](http://browsehappy.com/) of each browser at the time of writing is:
```
Chrome: 43
Firefox: 38
IE: 11
Opera: 30
Safari: 8
```

**Notice** The `size` field on the Set is no longer read-only. The reason for that is 25% time reduction (from 90-100 ms down to 15-25 ms) on the following test: http://jsfiddle.net/jto18oyh/2/. It is not excellent, but since a set is a core component and performance is important, I am in favor for this change

**Also:** Backward compatible 

---
#### Numbers
On the stacked plot test, the improvement is substantial. (if looking into the JSfiddle, I suggest open the console before clicking on the link)

Before http://jsfiddle.net/r44qLe06/2/
```
t: 11793.249ms
t: 12763.007ms
t: 13495.781ms
t: 14244.572ms
t: 14458.213ms
```
After http://jsfiddle.net/r44qLe06/1/
```
t: 1377.659ms
t: 1260.547ms
t: 1304.650ms
t: 1322.013ms
t: 1646.604ms
```


**QE Notes** Regression pass. IE should behave exactly the same. Firefox and Chrome should be faster everywhere as we improved the speed of our core Set and Map data structures. As a side note, stacking for stacked bar and stacked line should be improved by up to 10x