@bluong 2015-06-16
Having the `size` field not be readonly is incredibly tricky... as now we have a `Set` that can lie...