@aicioara 2015-06-16
I prefer that and performance, to a 4x slowdown tbh. Enforcing it to be read-only is trickier because a new object is created every time, hence garbage collection pressure