@bluong 2015-06-16
We should almost certainly submit a pull request to DefinitelyTyped as now we have a bunch of cruft in our codebase regarding any types.  I don't want to block this going through on that pull request, but having strong, clean is important enough that I would put that on high priority after this is done.