@jtlan 2015-06-16
There's probably definitions for `Set` and `Map` already.