@jtlan 2015-06-16
It seems like most of our large data-scale cases involve `number`s or `string`s. Can we update our shim to use a simple `Object` as a dictionary for those cases, with a `string`ified key? That would improve perf even in cases without the ES6 `Map` / `Set`.