@bluong 2015-06-16
@jtlan I wasn't able to find the `Map` / `Set` definitions in DefinitelyTyped, unfortunately :cry: