@aicioara 2015-06-16
@jtlan The IE issue is out of the scope of this PR, hence opened https://github.com/palantir/plottable/issues/2342

Addressed the other issue