@aicioara 2015-06-16
Actually removed those calls altogether as .forEach() is tested later