@jtlan 2015-06-17
@aicioara, can you explain what sort of testing might need to be done for this? It seems like the unit tests cover this pretty well right?