@crmorford 2015-06-17
Before this PR (on chrome) http://jsfiddle.net/v5p1d7vs/3/