@aicioara 2015-06-17
Performance testing: Stacking should be affected the most.
Regression pass: everything can be affected as we use the datastructures in lots of places