@crmorford 2015-06-17
Bar labels are in the wrong spots
![screen shot 2015-06-16 at 5 55 17 pm](https://cloud.githubusercontent.com/assets/7527733/8197730/ebfdc9ec-1450-11e5-8aa8-7863625810e5.png)