@crmorford 2015-06-17
Regression looks good!