@crmorford 2015-06-17
FF before http://jsfiddle.net/v5p1d7vs/8/
FF afer http://jsfiddle.net/v5p1d7vs/9/