@crmorford 2015-06-17
IE before (not as many data points) http://jsfiddle.net/v5p1d7vs/10/
IE after (about the same) http://jsfiddle.net/v5p1d7vs/11/